feat: env categories
