+++ .env.example
@@ -1,3 +1,4 @@
 REACT_APP_URL_CONTACTS="/contacts"
 REACT_APP_URL_MEMBERS="/members"
 REACT_APP_URL_ORGANIZATION="/organization"
+REACT_APP_URL_CATEGORIES="/categories"
